Split out some tests into a new test module
